+++ src/b3d-loader.ts
@@ -93,10 +93,10 @@ document.body.append(
 import * as BABYLON from '@babylonjs/core'
 import type { B3d } from './tosi-b3d'
 import {
-  conventionName,
   applyMaterialConventions,
   B3dChild,
   isOff,
+  isIgnored,
 } from './b3d-utils'
 import { DestroyableBehavior } from './destroyable-behavior'
 import type { CombatEvent, ChainLink } from './destroyable'
@@ -159,17 +159,17 @@ export class B3dLoader extends B3dChild {
         this.lights = lights
 
         for (const mesh of meshes) {
-          if (conventionName(mesh.name).includes('-ignore')) {
+          if (isIgnored(mesh.name)) {
             mesh.dispose()
           }
         }
         for (const node of transformNodes) {
-          if (conventionName(node.name).includes('-ignore')) {
+          if (isIgnored(node.name)) {
             node.dispose()
           }
         }
         for (const light of lights) {
-          if (conventionName(light.name).includes('-ignore')) {
+          if (isIgnored(light.name)) {
             light.dispose()
           } else if (
             light instanceof BABYLON.PointLight ||